[#4]/refactor : 중복 제거
